rename folder rali to rocAL
