+++ rocAL/rocAL_pybind/amd/rocAL/decoders.py
@@ -1,6 +1,6 @@
-import amd.rali.types as types
+import amd.rocAL.types as types
 import rali_pybind as b
-from amd.rali.pipeline import Pipeline
+from amd.rocAL.pipeline import Pipeline
 
 def image(*inputs, user_feature_key_map = None, path='', file_root ='', annotations_file= '', shard_id = 0, num_shards = 1, random_shuffle = False, affine=True, bytes_per_sample_hint=0, cache_batch_copy= True, cache_debug = False, cache_size = 0, cache_threshold = 0,
                  cache_type='', device_memory_padding=16777216, host_memory_padding=8388608, hybrid_huffman_threshold= 1000000, output_type = types.RGB,
