+++ rocAL/rocAL_pybind/amd/rocAL/fn.py
@@ -1,10 +1,10 @@
 import sys
-from amd.rali import readers
-from amd.rali import decoders
+from amd.rocAL import readers
+from amd.rocAL import decoders
 import inspect
-import amd.rali.types as types
+import amd.rocAL.types as types
 import rali_pybind as b
-from amd.rali.pipeline import Pipeline
+from amd.rocAL.pipeline import Pipeline
 
 
 def fish_eye(*inputs, device=None):
@@ -16,8 +16,7 @@ def fish_eye(*inputs, device=None):
 def brightness(*inputs, brightness=1.0, bytes_per_sample_hint=0, image_type=0,
                preserve=False, seed=-1, device=None):
 
-    kwargs_pybind = {"input_image0": inputs[0],
-                     "is_output": False, "alpha": None, "beta": None}
+    kwargs_pybind = {"input_image0": inputs[0], "is_output": False, "alpha": None, "beta": None}
     brightness_image = b.Brightness(Pipeline._current_pipeline._handle ,*(kwargs_pybind.values()))
     return (brightness_image)
 
@@ -26,7 +25,6 @@ def resize(*inputs, bytes_per_sample_hint=0, image_type=0, interp_type=1, mag_fi
             save_attrs=False, seed=1, temp_buffer_hint=0, device = None):
 
     # pybind call arguments
-    kwargs_pybind = {"input_image0": inputs[0], "dest_width": resize_x, "dest_height": resize_y,
-                     "is_output": False}
+    kwargs_pybind = {"input_image0": inputs[0], "dest_width": resize_x, "dest_height": resize_y, "is_output": False}
     resized_image = b.Resize(Pipeline._current_pipeline._handle ,*(kwargs_pybind.values()))
     return (resized_image)
