+++ rocAL/rocAL_pybind/amd/rocAL/plugin/pytorch.py
@@ -1,7 +1,7 @@
 import torch
 import numpy as np
 import rali_pybind as b
-import amd.rali.types as types
+import amd.rocAL.types as types
 
 class RALIGenericImageIterator(object):
     def __init__(self, pipeline):
@@ -29,14 +29,14 @@ class RALIGenericImageIterator(object):
 
         self.loader.copyImage(self.out_image)
         if((self.loader._name == "Caffe2ReaderDetection") or (self.loader._name == "CaffeReaderDetection")):
-         
+
             for i in range(self.bs):
                 size = b.getImageNameLen(self.loader._handle,i)
                 print(size)
                 self.array = np.array(["                 "])
-                
+
                 self.out=np.frombuffer(self.array, dtype=(self.array).dtype)
-            
+
                 b.getImageName(self.loader._handle, self.out ,i)
             return self.out_image ,self.out_bbox, self.out_tensor
         else:
@@ -97,11 +97,11 @@ class RALIGenericIterator(object):
             self.loader.copyToTensorNCHW(self.out, self.multiplier, self.offset, self.reverse_channels, int(self.tensor_dtype))
         else:
             self.loader.copyToTensorNHWC(self.out, self.multiplier, self.offset, self.reverse_channels, int(self.tensor_dtype))
-        
+
         if((self.loader._name == "Caffe2ReaderDetection") or (self.loader._name == "CaffeReaderDetection")):
             self.lis = []  # Empty list for bboxes
             self.lis_lab = []  # Empty list of labels
-            
+
             #Count of labels/ bboxes in a batch
             self.bboxes_label_count = np.zeros(self.bs, dtype="int32")
             self.count_batch = self.loader.GetBoundingBoxCount(self.bboxes_label_count)
@@ -114,20 +114,20 @@ class RALIGenericIterator(object):
             #Image sizes of a batch
             self.img_size = np.zeros((self.bs * 2),dtype = "int32")
             self.loader.GetImgSizes(self.img_size)
-            
+
             count =0
             sum_count=0
             for i in range(self.bs):
                 count = self.bboxes_label_count[i]
-  
+
                 self.label_2d_numpy = (self.labels[sum_count : sum_count+count])
                 self.label_2d_numpy = np.reshape(self.label_2d_numpy, (-1, 1)).tolist()
                 self.bb_2d_numpy = (self.bboxes[sum_count*4 : (sum_count+count)*4])
                 self.bb_2d_numpy = np.reshape(self.bb_2d_numpy, (-1, 4)).tolist()
-                
+
                 self.lis_lab.append(self.label_2d_numpy)
                 self.lis.append(self.bb_2d_numpy)
-                
+
                 sum_count = sum_count +count
 
             self.target = self.lis
