+++ rocAL/rocAL_pybind/example/caffe_reader.py
@@ -1,8 +1,8 @@
 import sys
-from amd.rali.plugin.pytorch import RALIClassificationIterator
-from amd.rali.pipeline import Pipeline
-import amd.rali.ops as ops
-import amd.rali.types as types
+from amd.rocAL.plugin.pytorch import RALIClassificationIterator
+from amd.rocAL.pipeline import Pipeline
+import amd.rocAL.ops as ops
+import amd.rocAL.types as types
 import os
 
 class HybridTrainPipe(Pipeline):
