+++ rocAL/rocAL_pybind/example/coco_pipeline.py
@@ -6,9 +6,9 @@ import torch
 import random
 import itertools
 
-from amd.rali.pipeline import Pipeline
-import amd.rali.ops as ops
-import amd.rali.types as types
+from amd.rocAL.pipeline import Pipeline
+import amd.rocAL.ops as ops
+import amd.rocAL.types as types
 import sys
 import numpy as np
