+++ rocAL/rocAL_pybind/example/new_api_test_file.py
@@ -2,11 +2,11 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 import random
-from amd.rali.plugin.pytorch import RALI_iterator
+from amd.rocAL.plugin.pytorch import RALI_iterator
 
-from amd.rali.pipeline import Pipeline
-import amd.rali.fn as fn
-import amd.rali.types as types
+from amd.rocAL.pipeline import Pipeline
+import amd.rocAL.fn as fn
+import amd.rocAL.types as types
 import sys
 import numpy as np
 import cv2
