+++ rocAL/rocAL_pybind/example/tf_classification_pipeline.py
@@ -1,14 +1,14 @@
-from amd.rali.plugin.tf import RALIIterator
-from amd.rali.pipeline import Pipeline
-import amd.rali.ops as ops
-import amd.rali.types as types
+from amd.rocAL.plugin.tf import RALIIterator
+from amd.rocAL.pipeline import Pipeline
+import amd.rocAL.ops as ops
+import amd.rocAL.types as types
 import sys
 import tensorflow as tf
 
 class HybridPipe(Pipeline):
 	def __init__(self, feature_key_map, tfrecordreader_type, batch_size, num_threads, device_id, data_dir, crop, oneHotLabels=0, rali_cpu = True):
 		super(HybridPipe, self).__init__(batch_size, num_threads, device_id, seed=12 + device_id,rali_cpu=rali_cpu)
-		self.input = ops.TFRecordReader(path=data_dir, index_path = "", reader_type=tfrecordreader_type, user_feature_key_map=feature_key_map, 
+		self.input = ops.TFRecordReader(path=data_dir, index_path = "", reader_type=tfrecordreader_type, user_feature_key_map=feature_key_map,
 			features={
 				'image/encoded':tf.FixedLenFeature((), tf.string, ""),
 				'image/class/label':tf.FixedLenFeature([1], tf.int64,  -1),
@@ -72,15 +72,15 @@ def main():
 
 	pipe = HybridPipe(feature_key_map=featureKeyMap, tfrecordreader_type=TFRecordReaderType, batch_size=bs, num_threads=nt, device_id=di, data_dir=imagePath, crop=cropSize, oneHotLabels=oneHotLabel, rali_cpu=raliCPU)
 	pipe.build()
-	
+
 	imageIterator = RALIIterator(pipe)
 	for i, (images_array, labels_array) in enumerate(imageIterator, 0):
 		print("\n\n",i)
 		print("\nIMAGES ARRAY:\n",images_array)
 		print("\nLABELS ARRAY:\n",labels_array)
 	imageIterator.reset()
-			
-		
+
+
 
 if __name__ == '__main__':
-    main() 
+    main()
