+++ rocAL/rocAL_pybind/example/tf_mnistTrainingExample/tf_mnist_classification_rali.py
@@ -1,9 +1,9 @@
 
 from __future__ import print_function
-from amd.rali.plugin.tf import RALIIterator
-from amd.rali.pipeline import Pipeline
-import amd.rali.ops as ops
-import amd.rali.types as types
+from amd.rocAL.plugin.tf import RALIIterator
+from amd.rocAL.pipeline import Pipeline
+import amd.rocAL.ops as ops
+import amd.rocAL.types as types
 import sys
 import tensorflow.compat.v1 as tf
 tf.disable_v2_behavior()
